Contract script to fetch data with cosign
In reality the Enterprise Contract won't see the pipeline run, or
even the cluster where the pipelinerun is. So this is a more
realistic script.

Make the task definition handle either PR_NAME or IMAGE_URL and
IMAGE_DIGEST.

(Note that the check will always fail currently because the rego
expects to see taskrun data present.)