Ensure data dir always exists
If the pipeline is missing or doesn't have the expect test result,
we get the following error and the task fails.

  find: '/shared/ec-work-dir/data': No such file or directory

The error is avoided if we ensure the data dir exists.

Create the policies dir as well for consistency  even though it
would be created anyway in lib/fetch/git.